deps/metric: move metric into pipe path and make metric pipe-special
